对UI的组件进行了抽象
