+++ lib/app/modules/diary/views/widgets/diary_editor.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/material.dart';
 import 'package:daily_satori/app_exports.dart';
 import 'package:daily_satori/app/styles/diary_style.dart';
 import 'package:image_picker/image_picker.dart';
